Use inline asm
This is less headache than using an external compiler.

For one, I can now build in Fedora instead of having to use distrobox.
Two, Cargo stopped giving me an accurate list of RV64 features randomly,
and now I don't have to deal with that

Disadvantage here is that debugging may be harder